feat: ping-3622 adjust move to anon flow
